requirements.txt
Flask
selenium
google-auth
google-auth-oauthlib
google-api-python-client
gunicorn
